Div work + Added d3.js SVG
Created a container div for the nav bar and main body. Allowed the main body to take up the remaining space on the screen. Created and SVG and added a visual(only a block) to figure out how to center it. Figured out how to center horizontally but got a headache trying to center vertically. Also fixed the scrolling/bottom margin issue. Thats all