Planning Work
created divs and styled them for the planning tab. created interactive buttons that can be clicked to add more courses to semester. Did this in order to utilize the same onclick function in the drag and drop feature in the future. Also messed up the nav bar, we need to fix it because we cannot scroll.