Delete Button
on hover a delete button shows up, you can click it and that div will be deleted. need to figure out how to create a div with the button in it when clicking the div, ask alberto how iframe works. Also when i click the trash can it also clicks the div for some reason so it makes a new div and looks like it did nothing but it did!